OCR/requirements.txt
Flask==2.2.5
Flask-Cors==5.0.0
opencv-python-headless==4.7.0.72
pillow>=10.3.0
ultralytics==8.3.76
pytesseract==0.3.10
numpy==1.24.3
easyocr==1.7.0
torch==2.5.1
torchvision==0.20.1
torchtext==0.18.0
pdf2image==1.16.3
gunicorn==20.1.0

gitpython>=3.1.30
setuptools>=70.0.0
pandas>=1.0.0
python-dotenv
